+++ src/nats_sinks/http/__init__.py
@@ -1,0 +1,31 @@
+# SPDX-FileCopyrightText: 2026 Johan Louwers <louwersj@gmail.com>
+# SPDX-License-Identifier: Apache-2.0
+
+"""First-party HTTP sink public API."""
+
+from nats_sinks.http.client import HttpClient, HttpRequest, HttpResponse, StandardHttpClient
+from nats_sinks.http.config import HttpIdempotencyConfig, HttpRetryConfig, HttpSinkConfig
+from nats_sinks.http.mapping import (
+    HttpPreparedRequestBody,
+    http_body_value,
+    http_envelope_value,
+    http_idempotency_key,
+    prepare_http_body,
+)
+from nats_sinks.http.sink import HttpSink
+
+__all__ = [
+    "HttpClient",
+    "HttpIdempotencyConfig",
+    "HttpPreparedRequestBody",
+    "HttpRequest",
+    "HttpResponse",
+    "HttpRetryConfig",
+    "HttpSink",
+    "HttpSinkConfig",
+    "StandardHttpClient",
+    "http_body_value",
+    "http_envelope_value",
+    "http_idempotency_key",
+    "prepare_http_body",
+]